// src/shared/enums/type-order-by.ts
export enum TypeOrderBy {
  PRICE_ASC = 'price_asc',
  PRICE_DESC = 'price_desc',
  PUBLISHED_AT_ASC = 'publishedAt_asc',
  PUBLISHED_AT_DESC = 'publishedAt_desc',
}

export enum SortOrder {
  ASC = 'asc',
  DESC = 'desc',
}

export enum ShortTypeOrderBy {
  PRICE = 'price',
  PUBLISHED_AT = 'publishedAt',
}